Fix eslint
