+++ client/public/assets/css/stake/style.scss
@@ -16,3 +16,7 @@
 		display: none;
 	}
 }
+
+.cd_btn_stake_cspr.hide {
+	display: none;
+}
