+++ client/src/components/Common/Layout/Modal/Row.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const ModalRow = ({ label, value, customClass }) => (
+	<div className={`cd_confirm_modal_row ${customClass}`}>
+		<span className="cd_confirm_modal_label">{label}</span>
+		<span className="cd_confirm_modal_value">{value}</span>
+	</div>
+);
+
+export default ModalRow;
