+++ client/src/components/Common/Layout/Modal/SingleCurrencyRow.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const SingleCurrencyRow = ({ amount }) => (
+	<div className="cd_confirm_modal_row_single">
+		<span className="cd_confirm_modal_value">{amount}</span>
+	</div>
+);
+
+export default SingleCurrencyRow;
