+++ client/src/components/Common/Layout/Stake/Table.jsx
@@ -1,35 +1,10 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { Table } from 'react-bootstrap';
-
 import { toFormattedNumber, displayNaN } from '../../../../helpers/format';
 import CommonAction from '../../Button/CommonAction';
-
-const EmptyDelegation = ({ isLoading }) => (
-	<div className="cd_empty_delegation_row row">
-		<div className="cd_empty_delegation_section position-relative">
-			<img src="assets/image/no-staking-icon.svg" alt="empty-cspr-stake" />
-			<div className="cd_empty_delegation_message">
-				{isLoading
-					? 'Loading validators'
-					: 'You do not have any delegations yet. Stake CSPR, earn rewards and help Capser become more secure!'}
-			</div>
-		</div>
-	</div>
-);
-
-const TableActions = ({ validator, unDelegateFunc, disableAction }) => {
-	return (
-		<>
-			<span
-				className={`cd_tbl_action_undelegate ${disableAction ? 'disabled' : ''}`}
-				onClick={() => unDelegateFunc(validator)}
-			>
-				Undelegate
-			</span>
-		</>
-	);
-};
+import EmptyDelegation from './EmptyDelegation';
+import TableActions from './TableActions';
 
 const StakingAccountListComponent = ({
 	stakingDeployList = [],
