+++ client/src/components/Stake/index.jsx
@@ -1,59 +1,19 @@
 import React, { useEffect, useState } from 'react';
-import { Button, Alert } from 'react-bootstrap';
 import { useDispatch, useSelector } from 'react-redux';
-
-import { EXPLORER_URL, CASPER_SYMBOL } from '../../constants/key';
 import HeadingModule from '../Common/Layout/HeadingComponent/Heading';
 import StakingAccountList from '../Common/Layout/Stake/Table';
-import DelegateForm from '../Common/Layout/Stake/DelegateForm';
 import { MessageModal } from '../Common/Layout/Modal/MessageModal';
-
 import { getMassagedUserDetails, getPublicKey } from '../../selectors/user';
 import { getCurrentPrice } from '../../selectors/price';
 import { getValidators } from '../../selectors/validator';
 import { fetchValidators } from '../../actions/stakeActions';
 import { getPendingStakes } from '../../selectors/stake';
 import { useStakeFromValidators } from '../hooks/useStakeDeploys';
-import UndelegateForm from '../Common/Layout/Stake/UnDelegateForm';
 import { isLoadingRequest } from '../../selectors/request';
-
-const UnlockSingerWarning = ({ title, message }) => (
-	<section className="cd_staking_page">
-		<HeadingModule name={title} />
-		<div className="cd_staking_capser_locked">
-			<div className="cd_main_message">
-				<Alert variant="danger">{message}</Alert>
-			</div>
-		</div>
-	</section>
-);
-
-const ConfirmingTransactionsInfo = (transactions) => {
-	if (!transactions || !transactions.length) {
-		return;
-	}
-
-	if (transactions.length === 1) {
-		return (
-			<Alert variant={'info'} show={!!transactions.length}>
-				Confirming transaction ...{' '}
-				<Alert.Link
-					rel="noopner noreferrer"
-					target="_blank"
-					href={`${EXPLORER_URL}/deploy/${transactions[0].deployHash}`}
-				>
-					View on explorer
-				</Alert.Link>
-			</Alert>
-		);
-	}
-
-	return (
-		<Alert variant={'info'} show={!!transactions.length}>
-			Confirming transactions ...{' '}
-		</Alert>
-	);
-};
+import ConfirmingTransactionsInfo from '../Common/Layout/Stake/ConfirmingTransactionsInfo';
+import UnlockSingerWarning from '../Common/Layout/Stake/UnlockSingerWarning';
+import StakeForm from '../Common/Layout/Stake/Form';
+import StakeButton from './Button';
 
 const Stake = () => {
 	const dispatch = useDispatch();
@@ -80,18 +40,17 @@ const Stake = () => {
 
 	// Function
 	const handleToggle = () => {
-		if (publicKey) {
-			setSend(!send);
-		} else {
+		if (!publicKey) {
 			setShowError(true);
+			return;
 		}
+		setSend(!send);
 	};
 
 	const handleUndelegateToggle = () => {
 		if (send) {
 			setDefaultValidator(null);
 		}
-
 		handleToggle();
 	};
 
@@ -127,34 +86,17 @@ const Stake = () => {
 				{ConfirmingTransactionsInfo(pendingStakes)}
 				<HeadingModule name={'Staking'} />
 				<div className={`cd_staking_component ${toggleStakingForm}`}>
-					{showStakeBtn && (
-						<div className="cd_send_currency_btn_text cd_btn_stake_cspr">
-							<Button className=" cd_send_currency_btn" onClick={handleToggle}>
-								Stake CSPR
-							</Button>
-						</div>
-					)}
+					<StakeButton show={showStakeBtn} handleToggle={handleToggle} />
 					<div className="cd_staking_form">
-						{defaultValidator ? (
-							<UndelegateForm
-								fromAddress={publicKey}
-								csprPrice={currentPrice}
-								handleToggle={handleUndelegateToggle}
-								balance={displayBalance}
-								tokenSymbol={CASPER_SYMBOL}
-								stakedValidator={defaultValidator}
-							/>
-						) : (
-							<DelegateForm
-								defaultValidator={defaultValidator}
-								validators={validators}
-								handleToggle={handleToggle}
-								fromAddress={publicKey}
-								csprPrice={currentPrice}
-								balance={displayBalance}
-								tokenSymbol={CASPER_SYMBOL}
-							/>
-						)}
+						<StakeForm
+							defaultValidator={defaultValidator}
+							publicKey={publicKey}
+							validators={validators}
+							currentPrice={currentPrice}
+							handleUndelegateToggle={handleUndelegateToggle}
+							displayBalance={displayBalance}
+							handleToggle={handleToggle}
+						/>
 					</div>
 					{stakingDeployList && stakingDeployList.length > 0 && (
 						<h3 className="cd_transaction_list_main_heading">Your Delegations</h3>
