+++ client/src/helpers/currency.js
@@ -1,6 +1,6 @@
 import { BigNumber } from '@ethersproject/bignumber';
-import { MOTE_RATE } from '../constants/key';
 import Big from 'big.js';
+import { MOTE_RATE } from '../constants/key';
 
 /**
  * /**
